split some cleanup functions into separate file set tope level status to Running when NERSC job is running
